chore: restructure stuff
