+++ packages/blade/src/components/Icons/CreditCardIcon/CreditCardIcon.stories.tsx
