+++ packages/blade/src/components/Icons/_Svg/Path/Path.d.ts
