+++ packages/blade/src/components/Icons/_Svg/Path/index.ts
