+++ packages/blade/src/components/Icons/_Svg/Svg/Svg.d.ts
