+++ packages/blade/src/components/Icons/_Svg/Svg/Svg.web.tsx
