+++ packages/blade/src/components/Icons/_Svg/index.ts
@@ -1,0 +1,2 @@
+export { default as Path } from './Path';
+export { default as Svg } from './Svg';
