+++ packages/blade/src/components/Icons/index.tsx
@@ -1,5 +1,3 @@
-export { default as Path } from './Path';
-export { default as Svg } from './Svg';
 export { default as CreditCardIcon } from './CreditCardIcon';
 
 export type IconSize = 'large' | 'medium' | 'small' | 'xlarge' | 'xsmall' | 'xxsmall';
